+++ CSK_Module_MultiDeepLearning/scripts/MultiDeepLearning_Processing.lua
@@ -145,6 +145,13 @@ local function handleOnNewImageProcessing(image)
 end
 Script.serveFunction("CSK_MultiDeepLearning.processImage"..deepLearningInstanceNumberString, handleOnNewImageProcessing, 'object:1:Image', 'bool:?,float:[?*],string:[?*]')
 
+--- Function only used to forward the content from events to the served function.
+--- This is only needed, as deregistering from the event would internally release the served function and would make it uncallable from external.
+---@param image Image Image to process
+local function tempHandleOnNewImageProcessing(image)
+  handleOnNewImageProcessing(image)
+end
+
 --- Function to handle updates of processing parameters from Controller
 ---@param deepLearningNo int Number of instance to update
 ---@param parameter string Parameter to update
@@ -157,14 +164,14 @@ local function handleOnNewImageProcessingParameter(deepLearningNo, parameter, va
     if parameter == 'registeredEvent' then
       _G.logger:fine(nameOfModule .. ": Register DNN instance " .. deepLearningInstanceNumberString .. " on event " .. value)
       if imageProcessingParams.registeredEvent ~= '' then
-        Script.deregister(imageProcessingParams.registeredEvent, handleOnNewImageProcessing)
+        Script.deregister(imageProcessingParams.registeredEvent, tempHandleOnNewImageProcessing)
       end
       imageProcessingParams.registeredEvent = value
-      Script.register(value, handleOnNewImageProcessing)
+      Script.register(value, tempHandleOnNewImageProcessing)
 
     elseif parameter == 'deregisterFromEvent' then
       _G.logger:fine(nameOfModule .. ": Deregister instance " .. deepLearningInstanceNumberString .. " from event.")
-      Script.deregister(imageProcessingParams.registeredEvent, handleOnNewImageProcessing)
+      Script.deregister(imageProcessingParams.registeredEvent, tempHandleOnNewImageProcessing)
       imageProcessingParams.registeredEvent = ''
 
     elseif parameter == 'fullModelPath' then